feat(.install): remove apt requirements
A lot of the code was written for Debian in mind. However times change and I'm looking to be using a Mac again, this will break most of the environment if I'm not careful. I'm changing things to support MacOS with Homebrew. TBH the workarounds and logic needed are annoying and I may just skip all of that and go with a pre-req of homebrew and a bunch of brew installs.